requirements.txt
google-auth-oauthlib
requests
tkcalendar
pytz
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
tkinter
win10toast
plyer